[Common] Split PIDResponse between TPC and TOF (#10119)
