+++ Common/DataModel/CMakeLists.txt
@@ -16,6 +16,9 @@ o2physics_add_header_only_library(DataModel
                                           FT0Corrected.h
                                           Multiplicity.h
                                           PIDResponse.h
+                                          PIDResponseITS.h
+                                          PIDResponseTOF.h
+                                          PIDResponseTPC.h
                                           CollisionAssociationTables.h
                                           TrackSelectionTables.h
                                           McCollisionExtra.h
