+++ Common/TableProducer/PID/pidTOFBase.h
@@ -24,6 +24,7 @@
 // O2Physics
 #include "PID/ParamBase.h"
 #include "PID/PIDTOF.h"
+#include "Common/DataModel/PIDResponseTOF.h"
 #include "Common/DataModel/PIDResponse.h"
 
 static constexpr int nSpecies = 9;
