+++ Common/TableProducer/PID/pidTPC.cxx
@@ -35,7 +35,7 @@
 #include "Framework/ASoAHelpers.h"
 #include "ReconstructionDataFormats/Track.h"
 #include "CCDB/CcdbApi.h"
-#include "Common/DataModel/PIDResponse.h"
+#include "Common/DataModel/PIDResponseTPC.h"
 #include "Common/Core/PID/TPCPIDResponse.h"
 #include "Framework/AnalysisDataModel.h"
 #include "Common/DataModel/Multiplicity.h"
@@ -458,7 +458,7 @@ struct tpcPid {
     if (flagFull)
       tableFull(expSigma, nSigma);
     if (flagTiny)
-      aod::pidutils::packInTable<aod::pidtpc_tiny::binning>(nSigma, tableTiny);
+      aod::pidtpc_tiny::binning::packInTable(nSigma, tableTiny);
   };
 
   void processStandard(Coll const& collisions, Trks const& tracks, aod::BCsWithTimestamps const& bcs)
